refactor: move prover agent config to prover client package
