+++ yarn-project/circuit-types/src/config.ts
@@ -1,2 +1,1 @@
-export { ProverAgentConfig, proverAgentConfigMappings } from './interfaces/prover-agent.js';
 export { SequencerConfig, AllowedElement, SequencerConfigSchema } from './interfaces/configs.js';
